feat(accordion): extract animating height into a react hook
